requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pydantic==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
requests==2.31.0
httpx==0.25.2
celery==5.3.4
redis==5.0.1
rapidfuzz==3.5.2
pandas==2.1.3
polars==0.19.19
pytest==7.4.3
pytest-asyncio==0.21.1
streamlit==1.29.0
plotly==5.17.0
plotly-express==0.4.1